+++ pakyow-assets/lib/pakyow/application/behavior/assets/externals.rb
@@ -26,9 +26,10 @@ module Pakyow
           apply_extension do
             after "boot", "fetch.assets.externals" do
               if config.assets.externals.pakyow
-                external_script :pakyow, "^1.0.0", package: "@pakyow/js", files: [
+                external_script :pakyow, "^1.1.0-alpha.1", package: "@pakyow/js", files: [
                   "dist/pakyow.js",
                   "dist/components/confirmable.js",
+                  "dist/components/devtools.js",
                   "dist/components/form.js",
                   "dist/components/freshener.js",
                   "dist/components/navigator.js",