+++ pakyow-js/__tests__/features/support/helpers/components.js
@@ -1,0 +1,14 @@
+require("mutationobserver-shim");
+global.MutationObserver = window.MutationObserver;
+
+afterEach(() => {
+  jest.clearAllMocks();
+});
+
+afterEach(() => {
+  while(pw.Component.instances.length > 0) {
+    pw.Component.instances.pop();
+  }
+
+  pw.Component.clearObserver();
+});