+++ pakyow-js/__tests__/features/support/helpers/setup.js
@@ -1,0 +1,1 @@
+global.pw = require("../../../../src/index");