+++ pakyow-js/CHANGELOG.md
@@ -1,5 +1,10 @@
 # UNRELEASED
 
+  * [added] Introduces the `devtools` component
+    * Automatically reloads the page when the code changes
+    * Click on mode to restart in development or prototype mode
+    * Gracefully handles the difference between uris in prototype and development modes
+
 # 1.0
 
   * Hello, Web