+++ pakyow-presenter/CHANGELOG.md
@@ -1,5 +1,7 @@
 # UNRELEASED
 
+  * [changed] Prototype bar is now devtools, available in development and prototype modes
+
 # 1.0
 
   * Hello, Web